% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_ternary_boundary.R
\name{add_ternary_boundary}
\alias{add_ternary_boundary}
\title{Draw boundary of ternary diagram}
\usage{
add_ternary_boundary(...)
}
\arguments{
\item{...}{Arguments to pass to \code{lines()}.}
}
\description{
Add lines to make triangle border around ternary diagram.
}
