% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_ternary_gridlines.R
\name{add_ternary_gridlines}
\alias{add_ternary_gridlines}
\title{Draw gridlines for ternary diagram}
\usage{
add_ternary_gridlines(
  at = c(0.25, 0.5, 0.75),
  overhang = 0.05,
  label.cex = 0.5,
  ...
)
}
\arguments{
\item{at}{The levels to plot.}

\item{overhang}{Amount by which line should extend beyond diagram edge.}

\item{label.cex}{Size of text labeling levels.}
}
\description{
Adds three sets of gridlines to a ternary diagram, one for each
vertex.
}
