% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_ternary_lines.R
\name{add_ternary_lines}
\alias{add_ternary_lines}
\alias{add_ternary_arrow}
\title{Draw lines and arrows in ternary space}
\usage{
add_ternary_lines(point.1, point.2, overhang = 0, ...)

add_ternary_arrow(point.1, point.2, length = 0.075, angle = 30, ...)
}
\arguments{
\item{point.1, point.2}{The start and end points, as triples; in the
election context.}

\item{overhang}{Amount by which line should be extended in each direction.
Useful for plotting gridlines that go past the edges of the simplex.}

\item{...}{Other arguments to be passed to \code{lines()} or
\code{arrows()}.}
}
\description{
Pass a pair of points in standard coordinates (x, y, z)
to draw lines or arrows in ternary/barycentric coordinates.
}
