% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_ternary_point.R
\name{add_ternary_point}
\alias{add_ternary_point}
\title{Add point on ternary diagram}
\usage{
add_ternary_point(point, x.offset = 0, y.offset = 0, ...)
}
\arguments{
\item{point}{The point, as a triple. Can be normalized to shares or not.}

\item{x.offset, y.offset}{Amount by which point should be displaced horizontally/vertically.}

\item{...}{Other parameters to pass to \code{points()}.}
}
\description{
Takes a point in standard 3D space and plots it in ternary space.
}
