% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_ternary_polygon.R
\name{add_ternary_polygon}
\alias{add_ternary_polygon}
\title{Add polygon to ternary diagram}
\usage{
add_ternary_polygon(point.mat, border = NULL, border.lwd = 1, col = NA)
}
\arguments{
\item{point.mat}{A matrix of points: three columns, one row per point.}

\item{border}{Color of border, or NA to omit border.}

\item{border.lwd}{Linewidth of border.}

\item{col}{Color to fill polygon.}
}
\description{
Takes matrix of points in standard 3D space (three columns) and plots the corresponding polygon in ternary space.
}
