% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_ternary_text.R
\name{add_ternary_text}
\alias{add_ternary_text}
\title{Add text to ternary diagram}
\usage{
add_ternary_text(point, labels, x.offset = 0, y.offset = 0, ...)
}
\arguments{
\item{point}{The point, as a triple. Can be normalized to shares or not.}

\item{labels}{Label for text.}

\item{x.offset}{Amount by which point should be displaced horizontally/vertically.}

\item{y.offset}{Amount by which point should be displaced horizontally/vertically.}

\item{...}{Other parameters to pass to \code{text()}.}
}
\description{
Takes point in standard 3D space and adds text in ternary space.
}
