% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/brexit_prefs.R
\docType{data}
\name{brexit_prefs}
\alias{brexit_prefs}
\title{Preferences about Brexit}
\format{
A named list with 6 elements, one for each ordering of the three alternatives. Names (e.g. "Remain > Deal > No deal") specify the ordering.
}
\source{
Table 1 of report prepared for YouGov by Ben Lauderdale: \url{https://d25d2506sfb94s.cloudfront.net/cumulus_uploads/document/24pikzv1u7/Ben\%20Lauderdale\%20Brexit\%20deal\%20MRP\%20report.pdf}
}
\usage{
brexit_prefs
}
\description{
Estimated proportion of UK likely voters holding each preference ordering over the options "Remain", "Deal", and "No deal" in November 2018.
}
\keyword{datasets}
