% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_result_to_vector_of_vote_shares.R
\name{convert_result_to_vector_of_vote_shares}
\alias{convert_result_to_vector_of_vote_shares}
\title{Extract vote share vector from election result}
\usage{
convert_result_to_vector_of_vote_shares(result)
}
\arguments{
\item{result}{An election result: a vector, matrix, or data.frame of vote shares/counts. See details.}
}
\value{
A vector of vote shares (normalized to sum to one) with the order \code{c(ab, ac, ba, bc, ca, cb, ax, bx, cx)}. Note that the incomplete ballots are put at the end: this is slightly annoying but this is how the code developed and it would be annoying to fix it.
}
\description{
Used by \code{votevizr} plotting functions to process the election result passed by the user. The user may pass the election result either as vote counts or vote shares, and these counts/shares may be organized as a vector, a matrix, or a data frame.
}
\details{
If a numeric vector is passed, it should be in the form \code{c(ab, ac, ax, ba, bc, bx, ca, cb, cx)} where e.g. \code{ab} is the count or share of ballots ranking A first and B second (implicitly, C last) and e.g. \code{ax} is the count or share of ballots ranking A first and not ranking any candidate second. If there are no incomplete ballots (AX, BX, CX), these may be omitted. (The function adds 0s in the appropriate places.)

If a matrix is passed, it should be in the form\tabular{lllll}{
    \tab A \tab B \tab C \tab X \cr
   A \tab 0 \tab ab \tab ac \tab ax \cr
   B \tab ba \tab 0 \tab bc \tab bx \cr
   C \tab ca \tab cb \tab 0 \tab cx \cr
}


If there are no incomplete ballots, the final column may also be omitted.

If a data.frame is passed, it should be in the form\tabular{lll}{
   first_choice \tab second_choice \tab count_or_share \cr
   A \tab B \tab 475 \cr
   A \tab C \tab 284 \cr
   A \tab X \tab 6 \cr
   B \tab A \tab 621 \cr
   ... \tab ... \tab ... \cr
}


If there are no incomplete ballots those rows may be omitted. The function sorts on the first and second columns and assumes the resulting third column is \code{c(ab, ac, ax, ba, ...)}. This means that, if you use symbols others than A, B, C, X, you should make sure that your "X" replacement comes last in a sort. It could be NA.
}
\examples{
convert_result_to_vector_of_vote_shares(c(20, 5, 8, 8, 10, 20)) # no incomplete ballots 
convert_result_to_vector_of_vote_shares(c(20, 5, 0, 8, 8, 0, 10, 20, 0)) # also no incomplete ballots
convert_result_to_vector_of_vote_shares(c(20, 5, 1, 8, 8, 2, 10, 20, 1)) # some incomplete ballots
result <- matrix(c(0, 20, 5, 1, 8, 0, 8, 2, 10, 20, 0, 1), nrow = 3, ncol = 4, byrow = T)
convert_result_to_vector_of_vote_shares(result)
cands <- c("A", "B", "C", "X")
cand_combos <- expand.grid(cands, cands)
colnames(cand_combos) <- c("first", "second")
cand_combos <- cand_combos[cand_combos$first != "X" & cand_combos$first != cand_combos$second,]
cand_combos$count <- sample(1:100, size = nrow(cand_combos))
convert_result_to_vector_of_vote_shares(cand_combos)  
}
