% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/first_preference_shares.R
\name{first_preference_shares}
\alias{first_preference_shares}
\title{Extract first-preference shares from an ordinal election result}
\usage{
first_preference_shares(result, split = "")
}
\arguments{
\item{result}{The election result specified as a named list,
with each name being candidate names in order, e.g.
"Breed_Leno_Kim" or "ABC" or "Remain > Deal > No deal"}

\item{split}{The character on which to split the orderings
to recover candidate names.}
}
\description{
Given counts/shares of ballot orderings in various
formats, returns a vector with the proportion of
ballots ranking each candidate first.
}
