% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/first_preference_win_regions.R
\name{first_preference_win_regions}
\alias{first_preference_win_regions}
\alias{positional_first_preference_win_regions}
\alias{rcv_first_preference_win_regions}
\alias{condorcet_first_preference_win_regions}
\title{Extract first-preference win regions for plotting}
\usage{
first_preference_win_regions(
  result,
  split = "",
  method = "plurality",
  if_cycle = "empty",
  s = 0.5
)

positional_first_preference_win_regions(result, split = "", s = 0.5)

rcv_first_preference_win_regions(result, split = "")

condorcet_first_preference_win_regions(result, split = "", if_cycle = "empty")
}
\arguments{
\item{result}{The election result specified as a named list
of counts or shares for each ballot type, where each name
is a (possibly incomplete) ordering of candidates.
Examples: \itemize{
\item \code{list("abc" = 20, "acb" = 4, "bac" = 7, "bca" = 10, "cab" = 8, "cba" = 4, "ab" = 4, "a" = 1)}
\item \code{list("Remain > Deal > No deal" = 20, "Remain > No deal > Deal" = 4, etc.)}
}}

\item{split}{The character by which to split elements of
the named list to recover the candidate names.}

\item{method}{The election method: one of \code{"positional", "Condorcet", "RCV", "plurality", "Borda", "antiplurality"}.}

\item{if_cycle}{If \code{method = "Condorcet"},
specifies how the winner is determined in the
event of a Condorcet cycle. The first-preference win regions
leave open the cyclic area unless \code{if_cycle} is:
\itemize{
\item \code{"kemeny"}: Kemeny-Young method, i.e. the winner
is the candidate who loses by the smallest margin
\item a candidate name (must match a candidate name in \code{result}):
choose a particular candidate in the event of a cycle
}}

\item{s}{If \code{method = "positional"}, specifies the
relative value of a second preference, where a first
preference is worth 1 and a third preference is worth 0.
Special cases include plurality (\code{s = 0}) and
Borda count (\code{s = .5}).}
}
\description{
In an ordinal voting system, a candidate's "first-preference
win region" identifies first-preference shares where that candidate wins the election,
given the pattern of lower preferences. Each function below take counts/shares of ballot orderings and returns
a data frame with the vertices of each candidate's
first-preference win region for a given voting system.
}
