% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_ternary_boundary.R
\name{geom_ternary_boundary}
\alias{geom_ternary_boundary}
\title{Draw boundary for a ternary diagram}
\usage{
geom_ternary_boundary(alpha = 1, col = "black")
}
\arguments{
\item{alpha}{Opacity of boundary lines.}

\item{col}{Color of boundary lines.}
}
\description{
Adds a triangular boundary to a ggplot object drawn in ternary coordinates. Useful when you don't want to plot the first-preference win regions.
}
