% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_ternary_gridlines.R
\name{geom_ternary_gridlines}
\alias{geom_ternary_gridlines}
\title{Draw gridlines on a ternary diagram}
\usage{
geom_ternary_gridlines(
  at = (1:3)/4,
  line_overhang = 0.05,
  text_overhang = 0.1,
  alpha = 0.75,
  col = "darkgray",
  size = 2
)
}
\arguments{
\item{at}{A vector of the points at which the gridlines should be drawn.}

\item{line_overhang}{By how much should the gridlines overhang the edge of the ternary triangle?}

\item{text_overhang}{How far from the edge of the ternary triangle should the text labels be?}

\item{alpha}{Opacity of gridlines.}

\item{col}{Color of gridlines.}

\item{size}{Size of text labels.}
}
\description{
Adds a layer to a ggplot object drawn in ternary coordinates.
}
