% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_majority_and_plurality_lines.R
\name{majority_and_plurality_lines}
\alias{majority_and_plurality_lines}
\alias{add_majority_win_lines}
\alias{add_plurality_tie_lines}
\alias{add_plurality_win_lines}
\title{Draw significant lines for majority and plurality}
\usage{
add_majority_win_lines(overhang = 0, ...)

add_plurality_tie_lines(overhang = 0, ...)

add_plurality_win_lines(overhang = 0, ...)
}
\arguments{
\item{overhang}{Amount by which lines should be extended beyond edge of ternary diagram.}
}
\description{
Add lines to a ternary plot showing where a candidate wins a majority/plurality or ties
}
