% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rcv.R
\name{plot.av.result}
\alias{plot.av.result}
\title{Plot result of RCV election}
\usage{
plot.av.result(
  v.vec,
  vertex.labels = c("A", "B", "C"),
  add.fp.result = T,
  fp.result.col = "black",
  fp.result.cex = 1,
  shading.cols = c("#E495A566", "#86B87566", "#7DB0DD66"),
  from.v.vec = NULL,
  secondary.line.col = "gray",
  secondary.line.lwd = 2,
  main = NULL,
  new = T,
  border = "black",
  border.lwd = 1,
  space = 0.1,
  clipped = F,
  clipped.x.range = c(1/4, 3/4),
  clipped.y.range = c(0, 1/2),
  draw.elimination.regions = F,
  draw.majority.thresholds = F
)
}
\arguments{
\item{v.vec}{A vector of ballot counts or shares of length 9 or 6. If candidates 
are A, B, and C, then the components indicate the count/share
of ballots ranking the candidates (AB, AC, BA, BC, CA, CB, A, B, C); if the vector is length 6 the last three are assumed to be zero. The vertices will be plotted with A on the botton left, B on the top, and C on the bottom right.}

\item{vertex.labels}{Labels for the vertices, in order bottom-left, top, bottom-right.}

\item{add.fp.result}{Include a dot for the observed shares of first-preferences, i.e. top rankings?}

\item{fp.result.col}{Color for first preference result.}

\item{fp.result.cex}{Size of first preference result.}

\item{shading.cols}{Colors for sharing first-preference win regions, in order bottom-left, top, bottom-right.}

\item{from.v.vec}{If provided, show two results on one plot.}

\item{secondary.line.col}{Color for guidelines.}

\item{secondary.line.lwd}{Line width of guidelines}

\item{main}{Main title for plot.}

\item{new}{Draw new plot? If F, adds to existing plot.
= T,}

\item{border}{Color of border of ternary diagram.}

\item{border.lwd}{Linewidth of diagram border.}

\item{space}{How far from vertices to put labels? Default .1, where 1 is distance between bottom left and bottom right vertices.}

\item{clipped}{Allows to focus on one part of figure.}

\item{clipped.x.range}{Vector of range of x to show (like xlim), in standard coordinates.}

\item{clipped.y.range}{Vector of range of y to show (like ylim), in standard coordinates.}

\item{draw.elimination.regions}{Draws lines to show first preference shares where each candidate would be eliminated.}

\item{draw.majority.thresholds}{Draws lines to show where each candidate would win a majority of first preference shares.}
}
\value{
Beyond making the plot, the funtion returns a list of three points (in standard coordinates), each of
    which is an interior corner of a first-preference win region 
    in the diagram, where a majority tie line intersects with a 
    plurality tie line. For example, \code{AB.tie.C.max} is the
    intersection of the AB majority tie line with either the BC
    plurality tie line or the AC majority tie line, depending 
    on the pattern of second preferences. These points can 
    be useful for augmenting the plot.
}
\description{
Given ballot counts, represent the result of a ranked-choice voting 
election (aka alternative vote, preferential voting, instant-runoff voting, STV with one winner) on ternary diagram.
}
\details{
The dot shows the proportion of top ranks each candidate 
received; the division of the triangle into win regions shows 
the proportion of top ranks each candidate would need to win the election
(given the pattern of lower rankings).
}
\examples{
plot.av.result(c(20, 3, 3,6, 7, 19))
plot.av.result(c(20, 3, 3,6, 7, 19, 4,6,2))
}
