% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/condorcet_methods.R
\name{plot.condorcet.result}
\alias{plot.condorcet.result}
\title{Plot result of Condorcet election}
\usage{
plot.condorcet.result(
  v.vec,
  from.v.vec = NULL,
  add.fp.result = T,
  fp.result.col = "black",
  fp.result.cex = 1,
  secondary.line.col = "gray",
  secondary.line.lwd = 2,
  vertex.labels = c("A", "B", "C"),
  shading.cols = c("#E495A566", "#86B87566", "#7DB0DD66"),
  main = NULL,
  new = T,
  border = "black",
  space = 0.1,
  clipped = F,
  clipped.x.range = c(1/4, 3/4),
  clipped.y.range = c(0, 1/2),
  draw.majority.tie.lines = F,
  in.cycle = "empty"
)
}
\arguments{
\item{v.vec}{A vector of ballot counts or shares of length 9 or 6. If candidates 
are A, B, and C, then the components indicate the count/share
of ballots ranking the candidates (AB, AC, BA, BC, CA, CB, A, B, C); if the vector is length 6 the last three are assumed to be zero. The vertices will be plotted with A on the botton left, B on the top, and C on the bottom right.}

\item{from.v.vec}{If provided, show two results on one plot.}

\item{add.fp.result}{Include a dot for the observed shares of first-preferences, i.e. top rankings?}

\item{fp.result.col}{Color for first preference result.}

\item{fp.result.cex}{Size of first preference result.}

\item{secondary.line.col}{Color for guidelines.}

\item{secondary.line.lwd}{Line width of guidelines}

\item{vertex.labels}{Labels for the vertices, in order bottom-left, top, bottom-right.}

\item{shading.cols}{Colors for sharing first-preference win regions, in order bottom-left, top, bottom-right.}

\item{main}{Main title for plot.}

\item{new}{Draw new plot? If F, adds to existing plot.
= T,}

\item{border}{Color of border of ternary diagram.}

\item{space}{How far from vertices to put labels? Default .1, where 1 is distance between bottom left and bottom right vertices.}

\item{clipped}{Allows to focus on one part of figure.}

\item{clipped.x.range}{Vector of range of x to show (like xlim), in standard coordinates.}

\item{clipped.y.range}{Vector of range of y to show (like ylim), in standard coordinates.}

\item{draw.majority.tie.lines}{Draw majority tie lines for each pair of candidates.}

\item{in.cycle}{Specify "empty", "kemeny", "A" (lower left candidate), "B" (top candidate), or "C" (lower right candidate).}
}
\description{
Given ballot counts, represent the result of a three-candidate 
Condorcet election on a ternary diagram. The dot shows the proportion of top ranks each candidate 
received; the division of the triangle into win regions shows 
the proportion of top ranks each candidate would need to win the election
(given the pattern of lower rankings).
}
\details{
A Condorcet method is one where voters rank candidates and, if there is a candidate who defeats all other candidates in pairwise competition, that candidate is the winner. (A defeats B in pairwise competition is A is ranked higher than B on a majority of ballots.) There may not be such a candidate: there could be a "Condorcet cycle" in which e.g. A defeats B, B defeats C, but C defeats A. There are several methods for choosing a winner in the event of such a cycle. For now the user can leave the cyclic region empty, fill it according to the Kemeny-Young method (the winner is the candidate with the smallest losing margin), or specify a candidate who wins in that region.
}
\examples{
v.vec <- c(25, 6, 6,12, 15, 10)
plot.condorcet.result(v.vec, in.cycle = "empty")
plot.condorcet.result(v.vec, in.cycle = "kemeny")
plot.condorcet.result(v.vec, in.cycle = "A")
plot.condorcet.result(v.vec, in.cycle = "A", draw.majority.tie.lines = T)
}
