% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_blank_standard.R
\name{plot_blank_standard}
\alias{plot_blank_standard}
\title{Draw empty canvas in standard coordinates}
\usage{
plot_blank_standard(
  xlim = c(0, 1),
  ylim = c(0, 1),
  padding = 0.1,
  offset = 0.05,
  vertex_labels = c("A", "B", "C"),
  main = NULL,
  margins = c(2, 2, 1, 1),
  vertex_labels_cex = 1
)
}
\arguments{
\item{xlim}{The horizontal coordinates of the bottom-left and bottom-right vertices, respectively.}

\item{ylim}{The vertical coordinates of the bottom vertcies and top vertex, respectively.}

\item{padding}{Amount of space to add around vertices.}

\item{offset}{Distance from vertices to vertex labels.}

\item{vertex_labels}{Labels for vertices, in (lower-left, top, lower-right) order.}

\item{main}{Label for plot.}

\item{vertex_labels_cex}{Character expansion for figure labels.}

\item{label.offset}{Distance from vertices to vertex labels.}

\item{vertex.labels}{Labels for vertices, in (lower-left, top, lower-right) order.}
}
\description{
Draw an empty diagram to be filled by other plotting commands.
}
