% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_blank_ternary.R
\name{plot_blank_ternary}
\alias{plot_blank_ternary}
\title{Draw empty ternary diagram}
\usage{
plot_blank_ternary(
  space = 0.1,
  label.offset = 0.05,
  vertex.labels = c("A", "B", "C"),
  main = NULL,
  xlim = c(0, 1),
  ylim = c(0, sqrt(3/4))
)
}
\arguments{
\item{space}{Amount of space to add around vertices.}

\item{label.offset}{Distance from vertices to vertex labels.}

\item{vertex.labels}{Labels for vertices, in (lower-left, top, lower-right) order.}

\item{main}{Label for plot.}

\item{xlim}{The horizontal coordinates of the bottom-left and bottom-right vertices, respectively.}

\item{ylim}{The vertical coordinates of the bottom vertcies and top vertex, respectively.}
}
\description{
Draw an empty ternary diagram to be filled by other plotting commands..
}
