% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_positional_result.R
\name{plot_positional_result}
\alias{plot_positional_result}
\alias{plot_plurality_result_v2}
\alias{plot_borda_result_v2}
\alias{plot_antiplurality_result_v2}
\alias{plot_positional_result_v2}
\title{Plot result of positional election}
\usage{
plot_plurality_result_v2(result, ...)

plot_borda_result_v2(result, ...)

plot_antiplurality_result_v2(result, ...)

plot_positional_result_v2(
  result,
  s = 0.5,
  show_first_pref_result = T,
  fp_result_col = "black",
  fp_result_cex = 1,
  shading_cols = c("#E5F5E099", "#A1D99B99", "#31A35499"),
  vertex_labels = c("A", "B", "C"),
  vertex_labels_cex = 1,
  offset = 0.1,
  margins = c(1, 1, 1, 1),
  main = NULL,
  suppress_plot = F,
  new = T,
  fp_win_region_border = "darkgray",
  padding = 0.1,
  xlim = c(0, 1),
  ylim = NULL,
  draw_positional_tie_lines = F,
  tie_line_col = "gray",
  ternary_coordinates = T
)
}
\arguments{
\item{result}{The result of the election to be plotted. There are three ways to specify a result:
\itemize{
\item As a vector of ballot counts/shares of length 6 or length 9.
If candidates are A, B, and C, then the components should be
listed as (length 6 case) \code{c(ab, ac, ba, bc, ca, cb)} or
(length 9 case) \code{c(ab, ac, ax, ba, bc, bx, ca, cb, cx)} where
\code{ab} indicates the count or share of ballots listing \code{a}
first and \code{b} second (implicitly \code{c} last) and
\code{ax} indicates the count of share of ballots listing \code{a}
first and no one second.
\item As a matrix with a row for each of the three valid first choices (e.g. A, B, C) and a column for each valid second choice (A, B, C, and X if incomplete ballots are permitted)
\item As a data frame in "long" format with the first and second columns identifying valid combinations of first and second choices (respectively) and a column of counts/shares} See documentation for \code{convert_result_to_vector_of_vote_shares()} for more detail on acceptable \code{result} input.}

\item{s}{The score (between 0 and 1) given to candidates for second rankings, where 1 is given to top rankings and 0 to bottom rankings. 0 is plurality, .5 is Borda, 1 is anti-plurality.}

\item{main}{Main title for the plot.}

\item{new}{Draw new plot? If F, adds to existing plot.
= T,}

\item{xlim}{Allows to focus on one part of figure. These are in standard coordinates.}

\item{ylim}{Allows to focus on one part of figure. These are in standard coordinates.}

\item{draw_positional_tie_lines}{Draw positional tie lines for each pair of candidates.}

\item{ternary_coordinates}{Logical: plot on ternary coordinates or standard coordinates?}
}
\value{
In addition to making the plot (if suppress_plot = F), the function returns a list consisting of a list of \code{positional_tie_line_endpoints} (one per pair of candidates) and the \code{intersection_point} of these tie lines.
}
\description{
Given ballot counts/shares, represent the result of a three-candidate
positional election on a ternary diagram. The dot shows the proportion of top ranks each candidate
received; the division of the triangle into win regions shows
the proportion of top ranks each candidate would need to win the election
(given the pattern of lower rankings).
}
\details{
A positional method is one where voters rank candidates and candidates are awarded points depending on the rankings they receive. Given three candidates, a top ranking receives one point, a bottom ranking receives no points, and a second ranking receives \code{s} points.

Special cases of positional methods are plurality (\code{s=0}), Borda count (\code{s=.5}), and anti-plurality (\code{s=1}). Each of these methods gets a specific plotting function, but in the background it just passes the appropriate  \code{s} argument to the more general method \code{plot_positional_result()}.
}
\examples{
out <- plot_positional_result_v2(result = c(20, 3, 8, 8, 8, 11), s = 2/5)
ternary_points(out$intersection_point, col = "red", cex = .5)
}
