% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/positional_methods.R
\name{plot_positional_results}
\alias{plot_positional_results}
\alias{plot_plurality_result}
\alias{plot_borda_result}
\alias{plot_antiplurality_result}
\alias{plot_positional_result}
\title{Plot result of positional election}
\usage{
plot_plurality_result(result, ...)

plot_borda_result(result, ...)

plot_antiplurality_result(result, ...)

plot_positional_result(
  result,
  positional.s = 0.5,
  add.fp.result = T,
  fp.result.col = "black",
  fp.result.cex = 1,
  shading.cols = c("#E5F5E099", "#A1D99B99", "#31A35499"),
  secondary.line.col = "gray",
  secondary.line.lwd = 2,
  vertex.labels = c("A", "B", "C"),
  main = NULL,
  plot_it = T,
  new = T,
  border = "black",
  space = 0.1,
  xlim = c(0, 1),
  ylim = c(0, sqrt(3/4)),
  draw.positional.tie.lines = F
)
}
\arguments{
\item{result}{The result of the election to be plotted. There are three ways to specify a result:
\itemize{
\item As a vector of ballot counts/shares of length 6 or length 9.
If candidates are A, B, and C, then the components should be
listed as (length 6 case) \code{c(ab, ac, ba, bc, ca, cb)} or
(length 9 case) \code{c(ab, ac, ax, ba, bc, bx, ca, cb, cx)} where
\code{ab} indicates the count or share of ballots listing \code{a}
first and \code{b} second (implicitly \code{c} last) and
\code{ax} indicates the count of share of ballots listing \code{a}
first and no one second.
\item As a matrix with a row for each of the three valid first choices (e.g. A, B, C) and a column for each valid second choice (A, B, C, and X if incomplete ballots are permitted)
\item As a data frame in "long" format with the first and second columns identifying valid combinations of first and second choices (respectively) and a column of counts/shares} See documentation for \code{convert_result_to_vector_of_vote_shares()} for more detail on acceptable \code{result} input.}

\item{positional.s}{The point value (between 0 and 1) given to candidates for second rankings, where 1 is given to top rankings and 0 to bottom rankings. 0 is plurality, .5 is Borda, 1 is anti-plurality.}

\item{add.fp.result}{Include a dot for the observed shares of first-preferences, i.e. top rankings?}

\item{fp.result.col}{Color for first preference result.}

\item{fp.result.cex}{Size of first preference result.}

\item{shading.cols}{Colors for sharing first-preference win regions, in order bottom-left, top, bottom-right.}

\item{secondary.line.col}{Color for guidelines.}

\item{secondary.line.lwd}{Line width of guidelines}

\item{vertex.labels}{Labels for the vertices, in order bottom-left, top, bottom-right.}

\item{main}{Main title for the plot.}

\item{new}{Draw new plot? If F, adds to existing plot.
= T,}

\item{border}{Color of border of ternary diagram.}

\item{space}{How far from vertices to put labels? Default .1, where 1 is distance between bottom left and bottom right vertices.}

\item{xlim}{Allows to focus on one part of figure. These are in standard coordinates.}

\item{ylim}{Allows to focus on one part of figure. These are in standard coordinates.}

\item{draw.positional.tie.lines}{Draw positional tie lines for each pair of candidates.}
}
\value{
In addition to making the plot (if plot_it = T), the function returns a list consisting of a list of \code{positional.tie.line.endpoints} (one per pair of candidates) and the \code{intersection.point} of these tie lines.
}
\description{
Given ballot counts/shares, represent the result of a three-candidate
positional election on a ternary diagram. The dot shows the proportion of top ranks each candidate
received; the division of the triangle into win regions shows
the proportion of top ranks each candidate would need to win the election
(given the pattern of lower rankings).
}
\details{
A positional method is one where voters rank candidates and candidates are awared points depending on the rankings they receive. Given three candidates, a top ranking receives one point, a bottom ranking receives no points, and a second ranking receives \code{s} points.

Special cases of positional methods are plurality (\code{s=0}), Borda count (\code{s=.5}), and anti-plurality (\code{s=1}). Each of these methods gets a specific plotting function, but in the background it just passes the appropriate  \code{positional.s} argument to the more general method \code{plot_positional_result()}.
}
\examples{
out <- plot_positional_result(result = c(20, 3, 8, 8, 8, 11), positional.s = 2/5)
add.ternary.point(out$intersection.point, col = "red", cex = .5)
}
