% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rcv.R
\name{plot_rcv_result}
\alias{plot_rcv_result}
\title{Plot result of RCV election}
\usage{
plot_rcv_result(
  result,
  vertex.labels = c("A", "B", "C"),
  add.fp.result = T,
  fp.result.col = "black",
  fp.result.cex = 1,
  shading.cols = c("#E495A566", "#86B87566", "#7DB0DD66"),
  secondary.line.col = "gray",
  secondary.line.lwd = 2,
  main = NULL,
  new = T,
  border = "black",
  border.lwd = 1,
  space = 0.1,
  xlim = c(0, 1),
  ylim = c(0, sqrt(3/4)),
  draw.elimination.regions = F,
  draw.majority.thresholds = F
)
}
\arguments{
\item{result}{The result of the election to be plotted. There are three ways to specify a result:
\itemize{
\item As a vector of ballot counts/shares of length 6 or length 9.
If candidates are A, B, and C, then the components should be
listed as (length 6 case) \code{c(ab, ac, ba, bc, ca, cb)} or
(length 9 case) \code{c(ab, ac, ax, ba, bc, bx, ca, cb, cx)} where
\code{ab} indicates the count or share of ballots listing \code{a}
first and \code{b} second (implicitly \code{c} last) and
\code{ax} indicates the count of share of ballots listing \code{a}
first and no one second.
\item As a matrix with a row for each of the three valid first choices (e.g. A, B, C) and a column for each valid second choice (A, B, C, and X if incomplete ballots are permitted)
\item As a data frame in "long" format with the first and second columns identifying valid combinations of first and second choices (respectively) and a column of counts/shares} See documentation for \code{convert_result_to_vector_of_vote_shares()} for more detail on acceptable \code{result} input.}

\item{vertex.labels}{Labels for the vertices, in order bottom-left, top, bottom-right.}

\item{add.fp.result}{Include a dot for the observed shares of first-preferences, i.e. top rankings?}

\item{fp.result.col}{Color for first preference result.}

\item{fp.result.cex}{Size of first preference result.}

\item{shading.cols}{Colors for sharing first-preference win regions, in order bottom-left, top, bottom-right.}

\item{secondary.line.col}{Color for guidelines.}

\item{secondary.line.lwd}{Line width of guidelines}

\item{main}{Main title for the plot.}

\item{new}{Draw new plot? If F, adds to existing plot.
= T,}

\item{border}{Color of border of ternary diagram.}

\item{border.lwd}{Linewidth of diagram border.}

\item{space}{How far from vertices to put labels? Default .1, where 1 is distance between bottom left and bottom right vertices.}

\item{xlim, ylim}{Allows to focus on one part of figure. These are in standard coordinates.}

\item{draw.elimination.regions}{Draws lines to show first preference shares where each candidate would be eliminated.}

\item{draw.majority.thresholds}{Draws lines to show where each candidate would win a majority of first preference shares.}
}
\value{
Beyond making the plot, the funtion returns a list of three points (in standard coordinates), each of
which is an interior corner of a first-preference win region
in the diagram, where a majority tie line intersects with a
plurality tie line. For example, \code{AB.tie.C.max} is the
intersection of the AB majority tie line with either the BC
plurality tie line or the AC majority tie line, depending
on the pattern of second preferences. These points can
be useful for augmenting the plot.
}
\description{
Given ballot counts, represent the result of a ranked-choice voting
election (aka alternative vote, preferential voting, instant-runoff voting, STV with one winner) on ternary diagram.
}
\details{
The dot shows the proportion of top ranks each candidate
received; the division of the triangle into win regions shows
the proportion of top ranks each candidate would need to win the election
(given the pattern of lower rankings).
}
\examples{
plot_rcv_result(c(20, 3, 3,6, 7, 19))
plot_rcv_result(c(20, 3, 3,6, 7, 19, 4,6,2))
}
