% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/positional_method_utilities.R
\name{positional_tie_line_utils}
\alias{positional_tie_line_utils}
\alias{from_point_in_xyz_form_v2}
\alias{to_point_in_xyz_form_v2}
\alias{point_mat_in_xyz_form_v2}
\alias{ternary_point_mats_from_p_vec_and_s_v2}
\title{Locate endpoints of positional tie lines}
\usage{
from_point_in_xyz_form_v2(pxy, pyx, s)

to_point_in_xyz_form_v2(pxy, pyx, pzx, pzy, s)

point_mat_in_xyz_form_v2(p_vec, s)

ternary_point_mats_from_p_vec_and_s_v2(p_vec, s)
}
\arguments{
\item{pxy, pyx, pzx, pzy}{Second preference shares. e.g. \code{pxy} is the proportion of ballots that rank \code{y} second, given that they rank \code{x} first.}

\item{s}{The score given to second-ranked candidates in the positional method. \code{s} = 0 corresponds to plurality, \code{s}=.5 corresponds to Borda count.}

\item{p_vec}{A vector of second preference shares. If candidates are A, B, and C (plotted at lower left, top, and lower right), then \code{p_vec} is (pAB, pAC, pBA, pBC, pCA, pCB).}
}
\description{
These functions are used to plot positional election results. They locate the points where positional tie lines intersect edges of the ternary diagram.
}
