% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qplot_votevizr.R
\name{qplot_votevizr}
\alias{qplot_votevizr}
\title{Visualize ordinal election result on ternary diagram}
\usage{
qplot_votevizr(
  result,
  split = "",
  method = "positional",
  if_cycle = "empty",
  s = 0.5,
  vertex_varnames = NULL,
  vertex_labels = NULL,
  label_offset = 0.05,
  padding = 0.1,
  show_gridlines = T,
  show_fp_result = T,
  fp_cex = 1
)
}
\arguments{
\item{result}{The election result specified as a named list
of counts or shares for each ballot type, where each name
is a (possibly incomplete) ordering of candidates.
Examples: \itemize{
\item \code{list("abc" = 20, "acb" = 4, "bac" = 7, "bca" = 10, "cab" = 8, "cba" = 4, "ab" = 4, "a" = 1)}
\item \code{list("Remain > Deal > No deal" = 20, "Remain > No deal > Deal" = 4, etc.)}
}}

\item{split}{The character by which to split elements of
the named list to recover the candidate names.}

\item{method}{The election method: one of \code{"positional", "Condorcet", "RCV", "plurality", "Borda", "antiplurality"}.}

\item{if_cycle}{Who is the winner in the event of a Condorcet
cycle? (Relevant if \code{method = "Condorcet"}.)
Provide a candidate name (must match names in the
\code{result}), or
\itemize{
\item \code{"empty"}: no winner specified
\item \code{"kemeny"}: the Kemeny-Young method, i.e. the winner
is the candidate with the smalling losing margin
\item \code{"RCV" or "IRV"}: use ranked-choice voting
\item \code{"antiplurality", "Borda", "plurality"}: use
specified positional method
\item \code{"positional"}: use generic positional method (must
also specify \code{s})
}}

\item{s}{The value of a second ranking. If
\code{method = "positional"}, this \code{s} is used to draw
first-preference win regions. If \code{method = "Condorcet"} and
\code{if_cycle = "positional"}, this \code{s} is used to
draw first-preference win regions within the cyclic triangle.}

\item{vertex_varnames}{A vector of candidate names (as characters) in
the order in which they are assigned to vertices of the ternary diagram:
lower-right, top, lower-left. They must match the candidate names
in \code{result} exactly. If \code{NULL}, they are taken from
\code{result}.}

\item{vertex_labels}{Vertex labels for plotting. If \code{NULL}, taken from \code{vertex_varnames}.}

\item{label_offset}{How far above/below the
vertices should we place the vertex labels?}

\item{padding}{Amount of padding around diagram. Make bigger to use
long \code{vertex_labels}.}

\item{show_gridlines}{Add gridlines to the ternary diagram?}

\item{show_fp_result}{If \code{TRUE}, put a dot
showing the first-preference shares in \code{result}.}

\item{fp_cex}{\code{cex} parameter for plotting the first preference
result.}
}
\description{
Given an ordinal election/poll result as a named list, produces
a ternary diagram showing first-preference win regions and
the observed first-preference result.
}
