% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ternary_translations.R
\name{ternary}
\alias{ternary}
\alias{simplex.x}
\alias{simplex.y}
\alias{simplex.xy}
\title{Convert counts/shares for three elements to barycentric coordinates}
\usage{
simplex.x(x)

simplex.y(x)

simplex.xy(x)
}
\arguments{
\item{x}{A triple representing votes or vote shares for three
candidates. Translated to correspond to lower left, lower right,
and top vertices of ternary diagram (respectively).}
}
\value{
The x coordinate (for \code{simplex.x} and \code{simplex.y})
or the x and y coordinates (for \code{simplex.xy}).
}
\description{
These functions perform the translation from standard
coordinates to barycentric coordinates.
}
