% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ternary_coords.R
\name{ternary_coords}
\alias{ternary_coords}
\title{Convert 3D coordinates to 2D ternary space}
\usage{
ternary_coords(x)
}
\arguments{
\item{x}{Vector, matrix, or data.frame/tibble of coordinates in 3D.}
}
\description{
Converts coordinate triples (supplied as length-3 vector, 3-column data.frame/tibble, or 3-column matrix) to two-column matrix. Coordinates assigned to lower-right, top, and lower-left, respectively. May be supplied as counts (non-normalized) or shares (normalized).
}
