% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ternary_lines.R
\name{ternary_lines}
\alias{ternary_lines}
\title{Draw lines in ternary space}
\usage{
ternary_lines(x, overhang = 0, ...)
}
\arguments{
\item{x}{A matrix, data.frame, or tibble of 3D coordinates.}

\item{overhang}{Amount by which line should be extended
beyond start and end points -- useful for plotting gridlines that go past the edges of the simplex.}

\item{...}{Other arguments to be passed to \code{lines()}.}
}
\description{
Pass a matrix of 3D coordinates \code{x} and draw a line connecting them.
}
