% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ternary_points.R
\name{ternary_points}
\alias{ternary_points}
\title{Add points to ternary diagram}
\usage{
ternary_points(x, ...)
}
\arguments{
\item{x}{The point(s) to plot, as a vector or as a three-column matrix/data.frame/tibble.}

\item{...}{Other parameters to pass to \code{points()}.}
}
\description{
Pass a vector of coordinates in 3D space, or a matrix/data.frame/tibble of 3D coordinates \code{x}, and add points to existing ternary plot.
}
