% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ternary_polygon.R
\name{ternary_polygon}
\alias{ternary_polygon}
\title{Add polygon to ternary diagram}
\usage{
ternary_polygon(x, ...)
}
\arguments{
\item{x}{A matrix/data.frame/tibble of points: three columns, one row per point.}

\item{...}{Arguments to pass to \code{polygon()}.}
}
\description{
Takes three-column matrix/data.frame/tibble of points in standard 3D space and plots the corresponding polygon in ternary space.
}
