% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ternary_text.R
\name{ternary_text}
\alias{ternary_text}
\title{Add text to ternary diagram}
\usage{
ternary_text(x, ...)
}
\arguments{
\item{x}{The point(s) to plot, as a vector or as a three-column matrix/data.frame/tibble.}

\item{...}{Other parameters to pass to \code{text()} (including \code{labels}).}
}
\description{
Takes point coordinates or matrix/data.frame/tibble of coordinates in standard 3D space and adds text in ternary space.
}
