% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/positional_methods.R
\name{plot_positional_results}
\alias{plot_positional_results}
\alias{plot.plurality.result}
\alias{plot.borda.result}
\alias{plot.antiplurality.result}
\alias{plot.positional.result}
\title{Plot result of positional election}
\usage{
plot.plurality.result(v.vec, ...)

plot.borda.result(v.vec, ...)

plot.antiplurality.result(v.vec, ...)

plot.positional.result(
  v.vec,
  positional.s = 0.5,
  add.fp.result = T,
  fp.result.col = "black",
  fp.result.cex = 1,
  shading.cols = c("#E495A566", "#86B87566", "#7DB0DD66"),
  secondary.line.col = "gray",
  secondary.line.lwd = 2,
  vertex.labels = c("A", "B", "C"),
  main = NULL,
  plot.it = T,
  new = T,
  border = "black",
  space = 0.1,
  clipped = F,
  clipped.x.range = c(1/4, 3/4),
  clipped.y.range = c(0, 1/2),
  draw.positional.tie.lines = F
)
}
\arguments{
\item{v.vec}{A vector of ballot counts or shares of length 9 or 6. If candidates 
are A, B, and C, then the components indicate the count/share
of ballots ranking the candidates (AB, AC, BA, BC, CA, CB, A, B, C); if the vector is length 6 the last three are assumed to be zero. The vertices will be plotted with A on the botton left, B on the top, and C on the bottom right.}

\item{positional.s}{The point value (between 0 and 1) given to candidates for second rankings, where 1 is given to top rankings and 0 to bottom rankings. 0 is plurality, .5 is Borda, 1 is anti-plurality.}

\item{add.fp.result}{Include a dot for the observed shares of first-preferences, i.e. top rankings?}

\item{fp.result.col}{Color for first preference result.}

\item{fp.result.cex}{Size of first preference result.}

\item{shading.cols}{Colors for sharing first-preference win regions, in order bottom-left, top, bottom-right.}

\item{secondary.line.col}{Color for guidelines.}

\item{secondary.line.lwd}{Line width of guidelines}

\item{vertex.labels}{Labels for the vertices, in order bottom-left, top, bottom-right.}

\item{main}{Main title for plot.}

\item{new}{Draw new plot? If F, adds to existing plot.
= T,}

\item{border}{Color of border of ternary diagram.}

\item{space}{How far from vertices to put labels? Default .1, where 1 is distance between bottom left and bottom right vertices.}

\item{clipped}{Allows to focus on one part of figure.}

\item{clipped.x.range}{Vector of range of x to show (like xlim), in standard coordinates.}

\item{clipped.y.range}{Vector of range of y to show (like ylim), in standard coordinates.}

\item{draw.positional.tie.lines}{Draw positional tie lines for each pair of candidates.}
}
\value{
In addition to making the plot (if plot.it = T), the function returns a list consisting of a list of \code{positional.tie.line.endpoints} (one per pair of candidates) and the \code{intersection.point} of these tie lines.
}
\description{
Given ballot counts, represent the result of a three-candidate 
positional election on a ternary diagram. The dot shows the proportion of top ranks each candidate 
received; the division of the triangle into win regions shows 
the proportion of top ranks each candidate would need to win the election
(given the pattern of lower rankings).
}
\details{
A positional method is one where voters rank candidates and candidates are awared points depending on the rankings they receive. Given three candidates, a top ranking receives one point, a bottom ranking receives no points, and a second ranking receives \code{s} points. 

Special cases of positional methods are plurality (\code{s=0}), Borda count (\code{s=.5}), and anti-plurality (\code{s=1}). Each of these methods gets a specific plotting function, but in the background it just passes the appropriate  \code{positional.s} argument to the more general method \code{plot.positional.result()}.
}
\examples{
out <- plot.positional.result(v.vec = c(20, 3, 8, 8, 8, 11), positional.s = 2/5)
add.ternary.point(out$intersection.point, col = "red", cex = .5)
}
